refactor(web): extract Title component
