+++ web/components/ChartTitle/index.ts
@@ -1,0 +1,1 @@
+export { default, HightText } from "./ChartTitle";
