+++ web/components/PlayerBioCard/PlayerBioCard.module.scss
@@ -26,6 +26,7 @@
         color: $primary-color;
         font-size: 24px;
         font-weight: 700;
+        font-family: "Roboto Condensed";
       }
       margin-bottom: 9px;
     }
