--- web/components/Text/index.ts
@@ -1,1 +1,0 @@
-export { default, HightText } from "./Text";
