+++ web/components/TimeOnIceChart/TimeOnIceChart.module.scss
@@ -22,13 +22,6 @@
     align-items: center;
   }
 
-  .title {
-    all: unset;
-    font-size: 1.5rem;
-    font-family: "Roboto Condensed";
-    font-weight: 700;
-  }
-
   @media only screen and (min-width: $screen-large) {
     .chartWrapper {
       flex: 1;
